requirements.txt
# NDX RSI Quant - dependencies (Python 3.9+)
# Data & calculation
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0

# Technical indicator verification (optional: install TA-Lib C library or use conda)
# TA-Lib>=0.4.0

# Config
PyYAML>=6.0

# Backtest (optional, for full backtest)
# backtrader>=1.9.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
